Safer handling of record ID handling
Previously, ignored whether there was an error when mapping the flow cell ID or instrument name to a `String`.
Now, it's wrapped within an if statement to only proceed if no error occured.

This will result in flow cell IDs and instrument names not being counted if there is an error.
The number of errors is not currently recorded.