Moving ID file and FASTQ readers outside of `FilterFastqIter` struct
Structs cannot be self-referential, and the current implementation of the previous and current IDs and records would cause a self-referential struct if the readers were also held there.
So the readers have been moved out into their own variable, stored by the filtering function, and are passed into the iterating struct.